model/requirements.txt
xgboost==0.90
pandas_profiling==2.3.0
pweave==0.30.3
wordcloud==1.5.0
lightgbm==2.2.3
catboost==0.17.3